LayoutTests/http/tests/xmlhttprequest/set-dangerous-headers-expected.txt

CONSOLE MESSAGE: line 12: Refused to set unsafe header "ACCEPT-CHARSET"
CONSOLE MESSAGE: line 13: Refused to set unsafe header "ACCEPT-ENCODING"
CONSOLE MESSAGE: line 14: Refused to set unsafe header "ACCESS-CONTROL-REQUEST-HEADERS"
CONSOLE MESSAGE: line 15: Refused to set unsafe header "ACCESS-CONTROL-REQUEST-METHOD"
CONSOLE MESSAGE: line 21: Refused to set unsafe header "CONNECTION"
CONSOLE MESSAGE: line 22: Refused to set unsafe header "CONTENT-LENGTH"
CONSOLE MESSAGE: line 23: Refused to set unsafe header "CONTENT-TRANSFER-ENCODING"
CONSOLE MESSAGE: line 24: Refused to set unsafe header "COOKIE"
CONSOLE MESSAGE: line 25: Refused to set unsafe header "COOKIE2"
CONSOLE MESSAGE: line 26: Refused to set unsafe header "DATE"
CONSOLE MESSAGE: line 27: Refused to set unsafe header "DNT"
CONSOLE MESSAGE: line 28: Refused to set unsafe header "EXPECT"
CONSOLE MESSAGE: line 29: Refused to set unsafe header "HOST"
CONSOLE MESSAGE: line 30: Refused to set unsafe header "KEEP-ALIVE"
CONSOLE MESSAGE: line 31: Refused to set unsafe header "ORIGIN"
CONSOLE MESSAGE: line 32: Refused to set unsafe header "REFERER"
CONSOLE MESSAGE: line 33: Refused to set unsafe header "TE"
CONSOLE MESSAGE: line 34: Refused to set unsafe header "TRAILER"
CONSOLE MESSAGE: line 35: Refused to set unsafe header "TRANSFER-ENCODING"
CONSOLE MESSAGE: line 36: Refused to set unsafe header "UPGRADE"
CONSOLE MESSAGE: line 37: Refused to set unsafe header "USER-AGENT"
CONSOLE MESSAGE: line 38: Refused to set unsafe header "VIA"
CONSOLE MESSAGE: line 40: Refused to set unsafe header "Proxy-"
CONSOLE MESSAGE: line 41: Refused to set unsafe header "Proxy-test"
CONSOLE MESSAGE: line 42: Refused to set unsafe header "PROXY-FOO"
CONSOLE MESSAGE: line 44: Refused to set unsafe header "Sec-"
CONSOLE MESSAGE: line 45: Refused to set unsafe header "Sec-test"
CONSOLE MESSAGE: line 46: Refused to set unsafe header "SEC-FOO"
Test that setRequestHeader cannot be used to alter security-sensitive headers.

SUCCESS
